Adding next homework.

+++ DataTypesAndVar/ConvertorForYears.cs
@@ -9,7 +9,7 @@ namespace DataTypesAndVar
             int years = int.Parse(Console.ReadLine());
 
             int days = years * 365;
-            int hours = days * 24;
+            long hours = days * 24;
             long minutes = hours * 60;
 
             Console.WriteLine($"{years} years = {days} days = {hours} hours = {minutes} minutes.");
